Fix random crashes when running around the world
Load a world
Run in one direction for a while (works best in plains biomes)
The game will eventually crash
The game crashed while attempting to update a wheat block. This happened when playing survival and running around the game. It seemed to be fairly easy to reproduce if you just run around the world.

The error:
forge Cannot get property IntegerProperty{name=age, clazz=class java.lang.Integer, values=[0, 1, 2, 3, 4, 5, 6, 7]} as it does not exist in Block{minecraft:air}

On debugging I found that the block state after the tick update was replacing the block with an air block. I guessed this was due to terrain generation generating it in the air and destroying it on the first tick update when it tries to grow the block. I decided to reload the world and go to the exact position shown in the debugger. Sure enough, the position was two blocks above a pond, and there was some grain on the ground indicating the block had been destroyed.

Basically, if a crop is generated in the air it will be destroyed during the tick update and replaced with an air block. In this case we don't want to check for disease/mutation as the crop no longer exists and it will cause Minecraft to crash when it tries to get the Age property.

The fix was to check that the new block state is not an air block after the parent update.